requirements.txt
asgiref==3.8.1
atomicwrites==1.4.1
attrs==25.3.0
colorama==0.4.6
Django==4.2
django-autocomplete-light==3.12.1
django-bootstrap5==22.2
django-filter==25.1
djangorestframework==3.14.0
flake8==5.0.4
flake8-docstrings==1.7.0
iniconfig==2.1.0
mccabe==0.7.0
packaging==24.2
pep8-naming==0.13.3
pillow==11.2.1
pluggy==1.0.0.dev0
py==1.11.0
pycodestyle==2.9.1
pydocstyle==6.3.0
pyflakes==2.5.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2025.2
snowballstemmer==2.2.0
sqlparse==0.5.3
toml==0.10.2
typing_extensions==4.13.2
tzdata==2025.2
